requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
sqlalchemy==2.0.23
aiohttp==3.9.3
numpy==1.24.3
torch==2.1.0
python-dotenv==1.0.0
twilio==8.10.0
python-multipart==0.0.6
pydantic==2.5.2
psycopg2-binary==2.9.9
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
